requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
mssql-django>=1.4.0
pyodbc>=5.1.0
python-decouple>=3.8
drf-yasg>=1.21.0
django-filter>=24.1
django-cors-headers>=4.3.0
gunicorn>=21.2.0
